quash all build warnings

+++ source/DM_Entity.cpp
@@ -21,9 +21,9 @@ void DM_Entity::EnsureBounds(DM_Game* pGame)
 	if (xTile < 0.0f)
 		xTile = 0.0f;
 	if (xTile > width - 1)
-		xTile = width - 1;
+		xTile = (float)width - 1;
 	if (yTile < 0.0f)
 		yTile = 0.0f;
 	if (yTile > height - 1)
-		yTile = height - 1;
+		yTile = (float)height - 1;
 }
